requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
wandb>=0.15.0
tqdm>=4.64.0
opencv-python>=4.6.0
facenet-pytorch>=2.5.0
datasets>=2.0.0
scipy>=1.7.0